LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/container-queries/anchored-fallback-position-area-any.tentative-expected.txt


FAIL @container anchored(fallback: <position-area>) querying 'any' assert_equals: expected "yes" but got ""
